Add authorization

+++ app/Http/Controllers/PinController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use Gate;
+use App\User;
 use App\Pin;
 use Illuminate\Http\Request;
 
@@ -9,6 +11,16 @@ use App\Http\Requests;
 
 class PinController extends Controller
 {
+  /**
+   * Constructs the controller
+   */
+  public function __construct () {
+    $this->middleware('auth', ['except' => [
+      'index',
+      'show',
+    ]]);
+  }
+
   /**
    * Display a listing of the resource.
    *
@@ -41,14 +53,12 @@ class PinController extends Controller
    */
   public function store(Request $request)
   {
-    $this->validate($request, [
-      'title' => 'required',
-      'description' => 'required',
-    ]);
+    $this->validatePin($request);
 
     $pin = new Pin;
     $pin->title = $request->title;
     $pin->description = $request->description;
+    $pin->user_id = $request->user()->id;
     $pin->save();
 
     return redirect('pin/' . $pin->id)
@@ -80,6 +90,10 @@ class PinController extends Controller
   {
     $pin = Pin::find($id);
 
+    if (Gate::denies('modify', $pin)) {
+      abort(403);
+    }
+
     return view('pin.edit', [
       'pin' => $pin,
     ]);
@@ -94,12 +108,14 @@ class PinController extends Controller
    */
   public function update(Request $request, $id)
   {
-    $this->validate($request, [
-      'title' => 'required',
-      'description' => 'required',
-    ]);
-
     $pin = Pin::find($id);
+
+    if (Gate::denies('modify', $pin)) {
+      abort(403);
+    }
+
+    $this->validatePin($request);
+
     $pin->title = $request->title;
     $pin->description = $request->description;
     $pin->save();
@@ -118,6 +134,10 @@ class PinController extends Controller
   {
     $pin = Pin::find($id);
 
+    if (Gate::denies('modify', $pin)) {
+      abort(403);
+    }
+
     $pin->delete();
 
     return redirect('/');
@@ -125,8 +145,8 @@ class PinController extends Controller
 
   private function validatePin ($request) {
     $this->validate($request, [
-      'title' => 'required',
-      'description' => 'required',
+      'title' => 'required|min:6',
+      'description' => 'required|min:6',
     ]);
   }
 }
